module-3/requirements.txt
ask-sdk-core
ask-sdk-s3-persistence-adapter